Linux support
Changed trimming of file paths and opening of files in default file explorers.
Opening of file in default file explorer is now done in separate task.
Refresh of non indexed directories in create index window is not done, when task is cancelled.
Null paths are not indexed.
Changed setter for mount of memory device.